vilidators refactor
