+++ packages/report-server/src/reportBuilder/fetch/functions/helpers/index.ts
@@ -1,0 +1,6 @@
+export {
+  validateDataGroups,
+  validateDataElementsForEvents,
+  validateDataElementsForAnalytics,
+  validateAggregations,
+} from './validators';
